backend/requirements.txt
Flask==1.0.2
Flask-Bootstrap==3.2.0.2
simplejson==3.6.0
Pillow==2.5.1
